merges NumberInputField back into NumberComponent
